Renamed rifle to riscv_config
