+++ README.rst
@@ -1,11 +1,11 @@
-RIFLE
+RISCV-Config
 ========
 
 RISC-V Feature Legalizer
 
 LICENSE: BSD-3 Clause.
 
-Latest Documentation of RIFLE:
+Latest Documentation of RISCV-Config:
 
   * `HTML <https://rifle.readthedocs.io/>`_
   * `PDF  <https://gitlab.com/incoresemi/rifle/-/jobs/artifacts/master/raw/RIFLE.pdf?job=doc>`_
