+++ docs/requirements.txt
@@ -6,7 +6,6 @@ chardet==3.0.4
 doc8==0.8.0
 docutils==0.14
 gitdb2==2.0.5
-GitPython==2.1.11
 idna==2.8
 imagesize==1.1.0
 Jinja2==2.10.1
