+++ docs/source/intro.rst
@@ -2,12 +2,12 @@
 Introduction
 ############
 
-**RIFLE** (RISC-V Feature Leagalizer) is a YAML based framework which can be used to validate the specifications of a RISC-V implementation against the RISC-V privileged and unprivileged ISA spec and generate standard specification yaml file. 
+**RISCV-Config** (RISCV Configuration Leagalizer) is a YAML based framework which can be used to validate the specifications of a RISC-V implementation against the RISC-V privileged and unprivileged ISA spec and generate standard specification yaml file. 
 
 **Caution**: This is still a work in progress and non-backward compatible changes are expected to happen. 
 
 For more information on the official RISC-V spec please visit: `RISC-V Specs <https://riscv.org/specifications/>`_
 
-RIFLE [`Repository <https://gitlab.com/incoresemi/rifle>`_]
+RISCV-Config [`Repository <https://github.com/riscv/riscv-config>`_]
