+++ riscv_config/schemas/schema_platform.yaml
