+++ setup.py
@@ -3,7 +3,7 @@ import codecs
 import pip
 from setuptools import setup, find_packages
 
-import rifle
+import riscv_config
 
 # Base directory of package
 here = os.path.abspath(os.path.dirname(__file__))
@@ -18,8 +18,8 @@ def read(*parts):
 with open("README.rst", "r") as fh:
     long_description = fh.read()
 
-setup(name="rifle",
-      version=rifle.__version__,
+setup(name="riscv_config",
+      version=riscv_config.__version__,
       description="RISC-V Featrue Legalizer by Incoresemi Ltd.",
       long_description=long_description,
       classifiers=[
@@ -27,16 +27,16 @@ setup(name="rifle",
           "License :: OSI Approved :: BSD License",
           "Development Status :: 4 - Beta"
       ],
-      url='https://gitlab.com/incoresemi/rifle',
+      url='https://github.com/riscv/riscv-config',
       author='InCore Semiconductors Pvt. Ltd.',
       author_email='neelgala@incoresemi.com',
       license='BSD-3-Clause',
       packages=find_packages(),
       install_package_data=True,
-      package_dir={'rifle': 'rifle/'},
-      package_data={'rifle': ['schemas/*']},
+      package_dir={'riscv_config': 'riscv_config/'},
+      package_data={'riscv_config': ['schemas/*']},
       install_requires=['Cerberus>=1.3.1', 'ruamel.yaml>=0.16.0'],
       python_requires=">=3.7.0",
       entry_points={
-          "console_scripts": ["rifle=rifle.main:main"],
+          "console_scripts": ["riscv-config=riscv_config.main:main"],
       })
